+++ src/helpers/extract-all-genres.ts
